Add nova-neutron&ovs connectors for multinode
This change adds the required neutron and ovs settings and packages
to compute nodes. Compute nodes need OVS packages but also some neutron
settings in nova.conf. ::neutron is included because it is needed
to provide some references for neutron::agents::ml2::ovs. neutron
is not actually running on the compute node however although some
packages are installed.